Renaming test permission package to permissions in order to match code package
