+++ server/src/test/java/keywhiz/service/permissions/AlwaysAllowDelegatingPermissionCheckTest.java
@@ -1,10 +1,7 @@
-package keywhiz.service.permission;
+package keywhiz.service.permissions;
 
 import com.codahale.metrics.MetricRegistry;
 import java.util.Objects;
-import keywhiz.service.permissions.Action;
-import keywhiz.service.permissions.AlwaysAllowDelegatingPermissionCheck;
-import keywhiz.service.permissions.PermissionCheck;
 import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
